+++ roles/lib/files/FWO.Api.Client/Data/NwObjectElement.cs
@@ -2,15 +2,15 @@
 {
     public class NwObjectElement
     {
-        public int ElemId { get; set; }
-        public int TaskId { get; set; }
+        public long ElemId { get; set; }
+        public long TaskId { get; set; }
         public Cidr Cidr { get; set; }
         public long? NetworkId { get; set; }
 
         public NwObjectElement()
         {}
 
-        public NwObjectElement(string cidrString, int taskId)
+        public NwObjectElement(string cidrString, long taskId)
         {
             Cidr = new Cidr(cidrString);
             TaskId = taskId;
