+++ roles/lib/files/FWO.Api.Client/Data/RequestApproval.cs
@@ -6,10 +6,10 @@ namespace FWO.Api.Data
     public class RequestApproval : RequestApprovalBase
     {
         [JsonProperty("id"), JsonPropertyName("id")]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [JsonProperty("task_id"), JsonPropertyName("task_id")]
-        public int TaskId { get; set; }
+        public long TaskId { get; set; }
 
 
         public RequestApproval()
