+++ roles/lib/files/FWO.Api.Client/Data/RequestElement.cs
@@ -6,10 +6,10 @@ namespace FWO.Api.Data
     public class RequestElement : ElementBase
     {
         [JsonProperty("id"), JsonPropertyName("id")]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [JsonProperty("task_id"), JsonPropertyName("task_id")]
-        public int TaskId { get; set; }
+        public long TaskId { get; set; }
 
         [JsonProperty("request_action"), JsonPropertyName("request_action")]
         public string RequestAction { get; set; } = FWO.Api.Data.RequestAction.create.ToString();
